Paranoid mode: Make paranoid mode more versatile/configurable
This commit adds a new [paranoid] section in bcfg2.conf. This allows you
to specify a backup path other than the default as well as set a maximum
number of copies to keep around. You are now allowed multiple backup
copies, which are marked with a timestamp. This should help to prevent
accidental deletion when running in paranoid mode multiple times. The
paranoid section currently looks something like this:

[paranoid]
path = /my/custom/backup/path
max_copies = 5

Signed-off-by: Sol Jerome <solj@ices.utexas.edu>

git-svn-id: https://svn.mcs.anl.gov/repos/bcfg/trunk/bcfg2@5345 ce84e21b-d406-0410-9b95-82705330c041