Gamin: aggressively drain gamin events from the gamin lib interface. (Patch from Michael McCallister)
Gamin doesn't do sufficient checking on the fd between the client and gam_server, so it can end up
in a deadlock. This patch insures that the fd has been drained on the client read side before
attempting to write to it. These events are buffered in the Gamin file interface until the Core
extracts and processes them. (Detailed explanation from Mike as well)

git-svn-id: https://svn.mcs.anl.gov/repos/bcfg/trunk/bcfg2@5386 ce84e21b-d406-0410-9b95-82705330c041