Enhanced bcfg2-info to implement a new "builddir" command.
This new command is similar to the "build" command, except that
instead of writing the complete config to a single file, it writes the
ConfigFile elements from the config into separate files in their
appropriate locations in a directory tree.  This is based on the
functionality (and implementation) found in the
tools/bcfg2-export-config script, which does not currently work due to
being out of sync with updates made to Bcfg2 library functions on
which it depends.



git-svn-id: https://svn.mcs.anl.gov/repos/bcfg/trunk/bcfg2@5479 ce84e21b-d406-0410-9b95-82705330c041