Implement support for M2Crypto on the bcfg2 client.
While Python 2.6's ssl module is the way of the future, support for
M2Crypto helps the Bcfg2 client run on machines with older versions of
Python.

This commit provides basic support.  More work remains to be done on
general SSL support within Bcfg2.  Reference the following
work-in-progress document (or its successors) for things to consider:

http://tools.ietf.org/html/draft-saintandre-tls-server-id-check-02

See also the following thread for other thoughts:

http://thread.gmane.org/gmane.comp.sysutils.bcfg2.devel/3563



git-svn-id: https://svn.mcs.anl.gov/repos/bcfg/trunk/bcfg2@5480 ce84e21b-d406-0410-9b95-82705330c041