Reports: Fix interaction display (Patch from Tim Laszlo) [bugfix]
From mailing list:

This has bugged me for a while..  not sure where it broke but when clicking
on an action on a node the details no longer display.  Every entry displayed
"This item does not currently exist on the host but is specified to exist in
the configuration.".

This is for the "/elements/bad/266499/" views.

Also the clients affected list was no longer being displayed.  I changed the
look up for item here to throw a 404 if it wasnt found.

Signed-off-by: Sol Jerome <solj@ices.utexas.edu>

git-svn-id: https://svn.mcs.anl.gov/repos/bcfg/trunk/bcfg2@5561 ce84e21b-d406-0410-9b95-82705330c041