Upstart: Add new upstart client tool
Due to the nature of the way Upstart handles service specification,
turning 'servicename' off and on can be done via a configuration file
located at /etc/init/<servicename>.conf. Enabling a disabled service can
be done by making sure that the Upstart configuration file and the
service are bundled together.

Signed-off-by: Sol Jerome <solj@ices.utexas.edu>