SSHbase: Generate known_hosts file in consistent order (Patch from Lee Loucks)
From Ticket #869:

Because the order of a python set is dependent on the order of the hash
of entries of that set, the get_skn method generates a entries in the
ssh_known_hosts with the names ordered according to their has instead of
a consistent (from the user point of view) order. Some entries are IP,
hostname and other entries are hostname, IP.

This patch corrects that.

Signed-off-by: Sol Jerome <solj@ices.utexas.edu>

git-svn-id: https://svn.mcs.anl.gov/repos/bcfg/trunk/bcfg2@5793 ce84e21b-d406-0410-9b95-82705330c041