APT: Add sources.list* to __important__ list Frame: Fix bug in handling of __important__ entries
This should allow people the easier option of configuring their APT
sources in the /etc/sources.list.d dirctory instead of just relying on
sources.list to be correct. This commit also fixes a bug where the
client was relying on a 'ConfigFile' entry tag even in the case where a
Path entry was present.

Signed-off-by: Sol Jerome <solj@ices.utexas.edu>