Reports: Fix default time zone settings (Resolves Ticket #957)
We were previously defaulting to America/Chicago instead of leaving this
option up to the user. Setting the default to None will cause Django to
go by the system time.

Note also that Django sets the os.environ['TZ'] variable when this is
not none, so we want to be careful not to set that if we don't have to.

Signed-off-by: Sol Jerome <sol.jerome@gmail.com>