NagiosGen: Dependency support (from Stéphane Graber <stgraber@ubuntu.com>)
From the patch description:
As we have several hundred machines in Bcfg2 at work and are using NagiosGen
to generate our Nagios configuration, the dependency tree was quite a mess.

The following patch adds support for a 'parents.xml' file in the NagiosGen
directory. This one contains a basic dependency tree like this:
<Depends>
   <Depend name='something' on='something-else'>
</Depends>

The plugin will then use this file to add 'parents' tag to nagiosgen.cfg
when possible.

It's being added when generating nagiosgen.cfg instead of when generating the
individual host entries for two reasons:
- To avoid having to wait for a run of bcfg2 on all machines when changing
  parents.xml
- To also support dynamic parents for manually added entries in NagiosGen
  (example: network equipement).