Change condition test for python version to work on Mac OS X 10.6.6 with Xcode 3.2.5.
The test using /bin/expr and && was not working for me. When I changed
it to this format, I was able to get M2Crypto-0.20.2-py2.5-macosx-10.5-ppc.egg
inserted into the package when python was 2.5 or less. I used two if
statements because I had trouble comparing floating point numbers with bash's
if.