Revert "bcfg2: Write cached config using encoding from Options.py"
This reverts commit 5d69ff7e966c4ffa911c78d11a6879b48e90aef8.

As the "rawconfig" variable now holds a UTF-8 encoded string, it would
have to be decoded before re-encoding it using a different encoding.
However, the cached configuration shouldn't be written using a non-UTF-8
encoding anyway, as "bcfg2 -f <configuration>" currently doesn't accept
any other encodings.  (If this is to be changed, the XML encoding
declaration of the configuration would have to be adjusted accordingly.)