Cfg: Add support for perms='inherit' (Ticket #642)
This feature allows you to use the on-disk permissions of the file in
the Cfg repository rather than specifying them using the traditional
means in info.xml.

Note that this only works for the octal permissions of the file on disk
since the owner/group may not exist on the destination machine.

Signed-off-by: Sol Jerome <sol.jerome@gmail.com>