Enhanced bcfg2-admin viz to allow output to be limited to one client.
By default, bcfg2-admin viz creates a diagram that shows the complete
contents of the repository: all Groups, Bundles, and (optionally)
Hosts/Clients. In a complicated configuration, this can be an
overwhelming amount of information.

This change adds an --only-client option that can be used to limit the
elements on the diagram to those that apply to the named host.