Add verification that version_info['micro'] is a single digit.
Because of restrictions in Mac OS X packaging, Bcfg2's
Micro and Minor version numbers must be combined into an
integer in the OS X package (IFMinorVersion attribute).
In order for this to work, the micro version must be an
integer, otherwise we'll run into cases where the·
IFMinorVersion will not be in sequence with the Bcfg2
Micro and Minor versions.