Rewrote DirectoryBacked so it handles files contained in an arbitrary directory layout.
Previously, DirectoryBacked (and as a result Bundler, Deps, Rules,
Base, Pkgmgr, and others) only recognized XML files contained in the
top-level plugin directory, for example:

        Deps/foo.xml
	Deps/subdir/foo.xml    # <--- Ignored
        Bundler/bar.xml
	Bundler/subdir/baz.xml # <--- Ignored

Now it can support the following as well:

        Deps/debian-lenny/foo.xml
        Deps/debian-squeeze/foo.xml
        Bundler/group-a/bar.xml
        Bundler/group-b/baz.xml

Note that the directories and filenames do not factor into the
semantic meaning of the configuration specification. The contents of
foo.xml must stand alone, as if they were in the same single-level
directory as before.

In the case of Deps, Rules, Pkgmgr, and Svcmgr, you must use Groups
and priorities within the XML files as needed to ensure that Bcfg2 can
correctly resolve the configuration for your hosts. For example, prior
to this change you would use a single file like the following:

    Deps/foo.xml:

        <Dependencies priority="0">
            <Group name="debian-lenny">
                <Package name="foo">
                    <Path name="/etc/foo.conf"/>
                </Package>
            </Group>
            <Group name="debian-squeeze">
                <Package name="foo">
                    <Path name="/etc/foo.conf"/>
                    <Path name="/etc/bar.conf"/>
                </Package>
            </Group>
        </Dependencies>

Now you can use a pair of files in separate directories like the
following. Note how the groups within each file prevent there from
being two sources for a single package:

    Deps/debian-lenny/foo.xml:

        <Dependencies priority="0">
            <Group name="debian-lenny">
                <Package name="foo">
                    <Path name="/etc/foo.conf"/>
                </Package>
            </Group>
        </Dependencies>

    Deps/debian-squeeze/foo.xml:

        <Dependencies priority="0">
            <Group name="debian-squeeze">
                <Package name="foo">
                    <Path name="/etc/foo.conf"/>
                    <Path name="/etc/bar.conf"/>
                </Package>
            </Group>
        </Dependencies>

In the case of Bundler, individual filenames must remain unique
throughout the directory hierarchy, and they must match the bundle
name.