POSIX: Honor the user's umask
The umask was set to zero in order to address the issue that mknod(2)'s
mode argument is modified by the process's umask.  However, this umask
setting also affected auto-created parent directories of configuration
entries: their permissions were set to `drwxrwxrwx'.  So, we now call
chmod(2) after mknod(2) instead of setting the umask to zero.