Avoid reporting Genshi bundles as missing
Genshi bundles were reported as missing in the configuration with
error messages like the following:

   Client CLIENTNAME configuration missing bundles: GBUNDLE1:GBUNDLE2

This seems to be caused by the Bundle name being derived from the
filename, which was assumed to end in a four character extension of
".xml". When the extension was actually ".genshi", the ".ge" portion
was left behind, and the code couldn't match up the requested bundles
with the generated bundles.