Improve error handling of Packages plugin.
Provide a descriptive error message that avoids logging a traceback.

Error out (instead of continuing) upon receiving an HTTP error or an
indication that the URL is malformed.

Only write data to the cache file if it was successfully retrieved
from the URL.

Make log messages for Packages plugin use consistent "Packages:" prefix.