Packages: Don't reverse the order of <Source>s
If a package is available in multiple repositories, the first one
specified in the sources.xml file should be used for dependency
resolution.  In the past, the Packages code had to reverse the internal
order of sources to achieve this effect; with the current code, this is
no longer the case.