Options: Inform user when unable to read bcfg2.conf
Reported by mattikus on IRC.

The real problem here is that we are glossing over a ton of potential
errors with an "except: pass" block. We need to redo this so that we
catch all the potential errors and pass them on to the user. The error
we were seeing in this case was caused by insufficient permissions to
read the bcfg2.conf file. There is an exception that is thrown, but it
was glossed over in the calling try/except block.

Signed-off-by: Sol Jerome <sol.jerome@gmail.com>