DBStats: Add unpruned entry reporting (#659)
Path type='directory' entries previously had no 'reason' associated with
their 'dirty' state.

This change adds a database field for unpruned entries which are present
in a pruned directory and displays that list in the web reporting
interface.

There is also a change made to the POSIX client tool which forces it to
display the extra entry list when running in interactive mode without
debugging enabled.

Signed-off-by: Sol Jerome <sol.jerome@gmail.com>