Chkconfig client tool:  fix service start result logic
When the bcfg2 client installs a service turning it 'on' with
chkconfig, it then starts the service with 'service foo start'.  If
all these succeed, it should return a True success value.

This fixes a bug where starting the service succeeds, but is
interpreted as failure, ultimately causing the host to be reported as
dirty.