Fix daemonization issues with BuiltinCore
When BuiltinCore daemonizes, the fam setup is destroyed, perhaps because
it has set up (although not run) threads.

This fix moves the _daemonize() code into the backends.  In the Builtin
backend, the core forks first thing in __init__().  In the
CherryPy backend, the core runs the cherrypy engine Daemonizer()
at the beginning of _run().