XMLFileBacked: fixed XInclude
Clear out "extras" when each event is handled to ensure that we start
fresh.  Failure to do so prevents the wildcard XInclude munging to
work, so the first time an event is processed everything is fine; but
when subsequent events are processed, _follow_xincludes() thinks that
each xinclude tag has already been munged and only removes it.