make chkconfig tool driver respect servicemode
The bcfg2 man page states that -s disabled should stop bcfg2 from
attempting to modify any services, but the Chkconfig driver (at least)
does start the service during the Install phase even with -s disabled.
This patch adds support to the Chkconfig driver for the servicemode
config parameter.  It still does chkconfig --add, which I think makes
sense to happen as part of configuration, but it does not attempt
to actually start the service.