MultiprocessingCore: Dispatch metadata cache expiration to children
When the broker in a multiprocessing configuration expires its
metadata cache (e.g., when probe data is received), it must dispatch
that expiration call to its children.

This also makes the protocol for communication between the broker and
its children into a real RPC protocol, so we can do even more stuff in
the future.