POSIX: Ignore permissions error on auto-created dirs
If the POSIX client tool is run as a non-root user, it is very
likely that the _set_perms() call in _makedirs() will fail because
it cannot set the owner of the newly-created directories.  This
causes _makedirs() to return False, which in turn causes
POSIXFile.install() to bail out early.

Applying the reasoning in <https://github.com/Bcfg2/bcfg2/pull/108>
the freebie directories created by _makedirs should have mode and
ownership  done on a best-effort basis.  If a user needs parent
directories to have a specific ownership and mode, then they should
specify that directory in their configuration.