Probes: rewrite for new caching system
Substantially rewrote the way the Probes plugin caches data.
Formerly, it was structured to assume that probe data was stored in
probed.xml, with the database feature very much a second-class
citizen; this adds a proper abstraction layer between the persistent
storage and the plugin (and its caches).

Also rewrote most Probes unit tests to actually be useful unit tests,
not implementation tests.