Reporting: have bcfg2-report-collector be explicit about detaching
Following the same logic as 360ba2e7, we should be explicit about
the need to detach the bcfg2-report-collector process.

Side note: this bit me because I was starting the bcfg2 server
processes via SSH, and the way python-daemon checks for being
started by inetd is to see if stdin is a socket. (??)