XMLFileBacked: Monitor XIncludes whenever a FAM is available
This monitors XIncluded files even if should_monitor=False, since the
object monitoring the base file will not monitor XIncludes.  This
ensures that XIncluded files are properly monitored whenever possible,
particularly for Bundler and Properties.

This is a partial backport of 5b66845 -- as much of a backport as is
possible without a module-level FAM object.