Tests: Fix tests after 9a6a231
The addition of the call to load_xml in 9a6a231 causes the test to
fail because load_xml() expects to read a clients.xml file.  The
actual actual open calls in write_xml are dummied out with Mock,
so no file is written, and thus cannot be read back.  This commit
dummies out the load_xml and adds some more asserts for good measure.