Correctly upgrade or downgrade yum packages
Formerly, yum did an 'update' to install the correct version of a
package, even if the desired package was older than the installed
package.  This is wrong; it needs to do a downgrade.  This changes it
to downgrade when the desired package is older, and upgrade if it is
newer.

There is still the possibility of upgrading a package that should be
downgraded if the desired package is only partially specified, but
this should be very rare.