Enable bcfg2-yum-helper to depsolve for arches incompatible with server
By default, the yum dependency resolver uses the host's architecture
to filter compatible packages.  This prevents dependency resolution
when the bcfg2 client's architecture is incompatible with the
server's.

This workaround checks the <Arch/> element for each of the client's yum
sources, and if they are all identical, passes that architecture to
bcfg2-yum-helper to override the default.

The rpmUtils.arch module may only be configured for a single
architecture.  If multiple architectures are configured in yum
sources, we don't know which one to pick, so use the default behavior
instead.