misc/bcfg2.spec Remove duplicate checks for systemd in RPM scriptlets
In the RPM scriptlets, there's first a check for fedora 18 or greater,
then if that's not true, a check for fedora 16 or greater.  Due to
some bug in how nested %if statements work in RPM scriptlets, the
second test is evaluating true even on non-Fedora systems, which is
leading to systemd commands being put in RHEL6 RPM scriptlets.

This change removes the second check.  If there needed to be a check
for versions of Fedora 16 and 17, they will no longer work, but since
neither of those are supported versions of Fedora, I suspect we don't
need to include logic for them.