Default to only (En|De)crypt vars that need it
For both Encrypting and Decrypting of Properties
files, we should by default only attempt to execute
on elements that have an "encrypted" attribute defined.

The code will already attempt to encrypt every element
if nothing in the current document matches this xpath,
which catches the case of a user trying to fully
encrypt a completely new properties file.

Conflicts:
	src/lib/Bcfg2/Server/Encryption.py