Implement _get_package_command and append _sysv_pkg_path attribute
Instead of doing a partially complete Install() method for SYSV,
implements a custom _get_package_command that will use the
_sysv_pkg_path attribute added by the pkgmogrify call.

This will allow the installs to complete.  Unfortunately, the
single-pass install will still fail if there are any packages with an
http:// URL.  The pkgadd invocation for 'device' sources doesn't take
multiple packages and the 'datastream' invocation doesn't handle packages
with an HTTP URL.  Finally, there is no reliable standard naming
convention for SYSV datastream files, so the simplename attribute is
re-used.

There is a known issue with this patch - if any packages specified in
the PackageList have an http url, the single-pass install will produce
an error like:

    Trying single pass package install for pkgtype sysv
    pkgadd: ERROR: Failure occurred with http(s) negotiation: <'Peername' doesn't match 'host' or no matching entry>
    pkgadd: ERROR: unable to download package datastream from <http://install1.d.stor.en.desres.deshaw.com/jumpstart10U10/packages>.
    Single Pass Failed

because the command that results isn't valid syntax for pkgadd.  A
workaround would be to add code to skip the single-pass install if any
packages had the simplename attribute, or by checking the url for the
presence of 'http'.  I'm not sure if that should be fixed or if this is
reasonable in this case.