Removed ndiff from POSIXFile, only attempt unified diff.
This removes the ndiff, which is a bit extraneous and can cause up to 30 second
timeouts on larger files with many changes. unified_diff is faster and
generally more applicable, and is already what is shown to the admin on
interactive mode. This can save up to 30 seconds per file, per run.

A future commit can take the resulting unified diff and recreate the ndiff
specifically for the reporting if that is desired.