FileUpload plugin
I had issues with FileProbes plugin, and missed the capability of genshi
templates, so I mixed up Bundler and FileProbes to have what I want.  It
is tested in 1.3.4 I don't know how version changes affect it. It
basically works like Bundler plugin. There are some not so nice errors,
and solutions in it. Sorry about that I'm not an expert in Python.  I
also made some extensions compared to the original one like handling
missing groups and users (the original version just threw an exception
in that case) Hope to be useful for someone, any comments and feedbacks
are welcome