YUM: Add options to enable and disable Yum plugins
Adds two options you can define:
* disabled_plugins: A comma-separated list of plugins to disable
* enabled_plugins: A comma-separated list of plugins to enable

This allows you to run bcfg2 with certain plugins enabled or disabled
when they're not set that way in the yum configuration.  This is
useful because the Bcfg2 YUM plugin is initialized before it can read
in any files that might overwrite yum plugin configuration.