Lint: fix default value for lint_plugins
This is a bit tricky because we need the enabled server plugins for the
default value. To have these available before parsing lint_plugins the
server plugins needs to be parsed during early parsing mode.

Additional there are lint plugins that has no matching server plugin.
To identify them, this adds a reference from the lint plutins to the
name of the matching server plugins in the member variable
__serverplugin__. If this variable is None the plugin is always used.

If a lint plugin could not be imported (maybe because of a missing
dependency) the plugin is ignored and not considered as a default value.