Bundler: Catch all exceptions in BuildStructures
This fixes a regression in 89e7afbf74ffb that made an unexpected
failure in a bundle (e.g., an arbitrary exception raised by a
templated bundle) prevent all subsequent bundles from being built,
too. Restoring the blanket 'except' makes it so that an unexpected
failure only affects the specific bundle that fails.